Remove AdaptMap Class
